move enums from tests to constants.js
