+++ test/constants.js
@@ -1,0 +1,21 @@
+const VoteType = {
+  Against: 0,
+  For: 1,
+  Abstain: 2,
+};
+
+const ProposalState = {
+  Pending: 0,
+  Active: 1,
+  Canceled: 2,
+  Defeated: 3,
+  Succeeded: 4,
+  Queued: 5,
+  Expired: 6,
+  Executed: 7,
+};
+
+module.exports = {
+  VoteType,
+  ProposalState,
+};
